more corrections on 0 create array

+++ 0x0B-malloc_free/0-create_array.c
@@ -9,13 +9,13 @@ char *create_array(unsigned int size, char c)
 {
 	if (size == 0)
 		return (NULL);
-	unsigned int i;
+	int i;
 	char *array;
 
 	array = (char *) malloc(sizeof(char) * size);
 	if (array == NULL)
 		return (NULL);
-	for (i = 0; i < size; i++)
+	for (i = 0; i < (int) size; i++)
 	{
 		array[i] = c;
 	}
